Finalizando a Service

+++ src/main/java/br/com/fiap/javaChallenge/repository/produto/NegociavelRespository.java
@@ -3,5 +3,5 @@ package br.com.fiap.javaChallenge.repository.produto;
 import br.com.fiap.javaChallenge.entity.produto.Negociavel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface NegociavelRespository extends JpaRepository<Negociavel, Integer> {
+public interface NegociavelRespository extends JpaRepository<Negociavel, Long> {
 }
